relocate common tests to tests/common.rs
